Always parenthesize commented return types
