+++ crates/ruff_python_formatter/src/expression/mod.rs
@@ -189,24 +189,38 @@ impl Format<PyFormatContext<'_>> for MaybeParenthesizeExpression<'_> {
         };
 
         match needs_parentheses {
-            OptionalParentheses::Multiline if *parenthesize != Parenthesize::IfRequired => {
-                if can_omit_optional_parentheses(expression, f.context()) {
-                    optional_parentheses(&expression.format().with_options(Parentheses::Never))
-                        .fmt(f)
-                } else {
+            OptionalParentheses::Multiline => match parenthesize {
+                Parenthesize::IfBreaksOrIfRequired => {
                     parenthesize_if_expands(&expression.format().with_options(Parentheses::Never))
                         .fmt(f)
                 }
-            }
+                Parenthesize::IfRequired => {
+                    expression.format().with_options(Parentheses::Never).fmt(f)
+                }
+                Parenthesize::Optional | Parenthesize::IfBreaks => {
+                    if can_omit_optional_parentheses(expression, f.context()) {
+                        optional_parentheses(&expression.format().with_options(Parentheses::Never))
+                            .fmt(f)
+                    } else {
+                        parenthesize_if_expands(
+                            &expression.format().with_options(Parentheses::Never),
+                        )
+                        .fmt(f)
+                    }
+                }
+            },
+            OptionalParentheses::Never => match parenthesize {
+                Parenthesize::IfBreaksOrIfRequired => {
+                    parenthesize_if_expands(&expression.format().with_options(Parentheses::Never))
+                        .fmt(f)
+                }
+
+                Parenthesize::Optional | Parenthesize::IfBreaks | Parenthesize::IfRequired => {
+                    expression.format().with_options(Parentheses::Never).fmt(f)
+                }
+            },
             OptionalParentheses::Always => {
                 expression.format().with_options(Parentheses::Always).fmt(f)
-            }
-            OptionalParentheses::Never if *parenthesize == Parenthesize::IfBreaksOrIfRequired => {
-                parenthesize_if_expands(&expression.format().with_options(Parentheses::Never))
-                    .fmt(f)
-            }
-            OptionalParentheses::Never | OptionalParentheses::Multiline => {
-                expression.format().with_options(Parentheses::Never).fmt(f)
             }
         }
     }
